requirements.txt
av==10.0.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
imageio-ffmpeg==0.4.8
PySocks==1.7.1
requests==2.31.0
scrapetube==2.4.0
soupsieve==2.4.1
typing_extensions==4.7.0
urllib3==2.0.3
youtube-dl @ file:///Users/seg/youtube-dl
